requirements.txt
Flask
requests
python-dotenv
python-pptx